requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
distro==1.9.0
fastapi==0.110.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-cli==0.0.21
langchain-community==0.0.32
langchain-core==0.1.42
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langserve==0.0.51
langsmith==0.1.47
markdown-it-py==3.0.0
marshmallow==3.21.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.17.1
orjson==3.10.0
packaging==23.2
# Editable Git install with no remote (pirate-speak==0.0.1)
-e d:\interview_projects\langchain-serve\pirate-speak\my-app\packages\pirate-speak
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rich==13.7.1
setuptools==68.2.2
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
sse-starlette==1.8.2
starlette==0.37.2
tenacity==8.2.3
tiktoken==0.6.0
tomlkit==0.12.4
tqdm==4.66.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.23.2
wheel==0.41.2
yarl==1.9.4
